+++ src/components/playlist-container.jsx
@@ -1,4 +1,5 @@
 import { useWatchLater } from "contexts";
+import { Link } from "react-router-dom";
 import "styles/playlist-container.css";
 
 export const PlaylistContainer = ({ video }) => {
@@ -29,13 +30,18 @@ export const PlaylistContainer = ({ video }) => {
           </div>
         </section>
 
-        <i class="cursor-pointer fas fa-times" onClick={() => deleteWatchLater(video)}></i>
+        <i
+          class="cursor-pointer fas fa-times"
+          onClick={() => deleteWatchLater(video)}
+        ></i>
 
-        <img
-          class="img"
-          src="https://media.istockphoto.com/photos/play-icon-youtube-picture-id1344290509?b=1&k=20&m=1344290509&s=170667a&w=0&h=nsr6-eek2_1H4OqmX5tdJE9LFVn20puWnO4xXx9j18g="
-          alt="video-poster"
-        />
+        <Link to={`/videoPlayer/${video.id}`}>
+          <img
+            class="img"
+            src="https://media.istockphoto.com/photos/play-icon-youtube-picture-id1344290509?b=1&k=20&m=1344290509&s=170667a&w=0&h=nsr6-eek2_1H4OqmX5tdJE9LFVn20puWnO4xXx9j18g="
+            alt="video-poster"
+          />
+        </Link>
       </div>
     </section>
   );
