+++ src/utils/index.js
@@ -1,3 +1,4 @@
 import { getFilteredVideos } from "./video-listing";
+import {addWatchLaterVideos,deleteWatchLaterVideos} from "./watch-later";
 
-export { getFilteredVideos };
+export { getFilteredVideos,addWatchLaterVideos,deleteWatchLaterVideos };
